[backend/drive] Switch to stream processing for remote media
This makes sure that files larger than the configured maximum remote media cache size are not loaded into memory (if the size is known), or are only loaded into memory until the configured maximum size before getting discarded (if the size is not known)

+++ Iceshrimp.Backend/Core/Configuration/Config.cs
@@ -100,8 +100,8 @@ public sealed class Config
 
 	public sealed class StorageSection
 	{
-		public readonly int?      MaxCacheSizeBytes;
-		public readonly int?      MaxUploadSizeBytes;
+		public readonly long?      MaxCacheSizeBytes;
+		public readonly long?      MaxUploadSizeBytes;
 		public readonly TimeSpan? MediaRetentionTimeSpan;
 
 		public bool              CleanAvatars = false;
@@ -170,10 +170,11 @@ public sealed class Config
 
 				MaxUploadSizeBytes = suffix switch
 				{
-					null => num,
-					'k' or 'K' => num * 1024,
-					'm' or 'M' => num * 1024 * 1024,
-					'g' or 'G' => num * 1024 * 1024 * 1024,
+					null       => num,
+					'k' or 'K' => num * 1024L,
+					'm' or 'M' => num * 1024L * 1024,
+					'g' or 'G' => num * 1024L * 1024 * 1024,
+
 					_ => throw new Exception("Unsupported suffix, use one of: [K]ilobytes, [M]egabytes, [G]igabytes")
 				};
 			}
@@ -200,10 +201,11 @@ public sealed class Config
 
 				MaxCacheSizeBytes = suffix switch
 				{
-					null => num,
-					'k' or 'K' => num * 1024,
-					'm' or 'M' => num * 1024 * 1024,
-					'g' or 'G' => num * 1024 * 1024 * 1024,
+					null       => num,
+					'k' or 'K' => num * 1024L,
+					'm' or 'M' => num * 1024L * 1024,
+					'g' or 'G' => num * 1024L * 1024 * 1024,
+
 					_ => throw new Exception("Unsupported suffix, use one of: [K]ilobytes, [M]egabytes, [G]igabytes")
 				};
 			}
